+ Optimized drawing functions by removing redundant operations. + Added slow motion feature to ced_demo_12.

+++ ced/ced_draw.h
@@ -26,7 +26,7 @@ namespace ced
 		, ::ced::view_grid<uint32_t>					depthBuffer
 		);
 	int								drawTriangle		(::ced::view_grid<::ced::SColorBGRA> pixels	, ::ced::STriangle2	<int32_t>	triangle	, ::ced::SColorBGRA color);
-	int								drawTriangle		(::ced::SCoord2<uint32_t> targetSize	, ::ced::STriangle3	<float>		triangle	, ::ced::container<::ced::SCoord2<int32_t>> & pixelCoords, ::ced::container<::ced::STriangleWeights<double>> & proportions, ::ced::view_grid<uint32_t> depthBuffer);
+	int								drawTriangle		(::ced::SCoord2<uint32_t> targetSize	, ::ced::STriangle3	<float>		triangle	, ::ced::container<::ced::SCoord2<int32_t>> & pixelCoords, ::ced::container<::ced::STriangleWeights<float>> & proportions, ::ced::view_grid<uint32_t> depthBuffer);
 	int								drawQuadTriangle
 		( const ::ced::view_grid<::ced::SColorBGRA>			targetPixels
 		, const ::ced::SGeometryQuads						& geometry
@@ -36,7 +36,7 @@ namespace ced
 		, const ::ced::SCoord3<float>						& lightVector
 		, const ::ced::SColorBGRA							color
 		, ::ced::container<::ced::SCoord2<int32_t>>			& pixelCoords
-		, ::ced::container<::ced::STriangleWeights<double>>	& pixelVertexWeights
+		, ::ced::container<::ced::STriangleWeights<float>>	& pixelVertexWeights
 		, ::ced::view_grid<uint32_t>						depthBuffer
 		);
 	int								drawQuadTriangle
@@ -47,7 +47,7 @@ namespace ced
 		, const ::ced::SMatrix4<float>						& matrixTransformView
 		, const ::ced::SCoord3<float>						& lightVector
 		, ::ced::container<::ced::SCoord2<int32_t>>			& pixelCoords
-		, ::ced::container<::ced::STriangleWeights<double>>	& pixelVertexWeights
+		, ::ced::container<::ced::STriangleWeights<float>>	& pixelVertexWeights
 		, ::ced::view_grid<::ced::SColorBGRA>				textureImage
 		, ::ced::container<::ced::SCoord3<float>>			& lightPoints
 		, ::ced::container<::ced::SColorBGRA>				& lightColors
@@ -63,7 +63,7 @@ namespace ced
 		, const ::ced::SMatrix4<float>						& matrixTransformView
 		, const ::ced::SCoord3<float>						& lightVector
 		, ::ced::container<::ced::SCoord2<int32_t>>			& pixelCoords
-		, ::ced::container<::ced::STriangleWeights<double>>	& pixelVertexWeights
+		, ::ced::container<::ced::STriangleWeights<float>>	& pixelVertexWeights
 		, ::ced::view_grid<::ced::SColorBGRA>				textureImage
 		, ::ced::container<::ced::SCoord3<float>>			& lightPoints
 		, ::ced::container<::ced::SColorBGRA>				& lightColors
@@ -77,7 +77,7 @@ namespace ced
 		, const ::ced::STriangle2	<float>					& triangleTexCoords
 		, const ::ced::SCoord3		<float>					& lightVector
 		, ::ced::container<::ced::SCoord2<int32_t>>			& pixelCoords
-		, ::ced::container<::ced::STriangleWeights<double>>	& pixelVertexWeights
+		, ::ced::container<::ced::STriangleWeights<float>>	& pixelVertexWeights
 		, ::ced::view_grid<::ced::SColorBGRA>				textureImage
 		, ::ced::container<::ced::SCoord3<float>>			& lightPoints
 		, ::ced::container<::ced::SColorBGRA>				& lightColors
@@ -92,7 +92,7 @@ namespace ced
 		, const ::ced::SCoord3<float>						& lightVector
 		, ::ced::SColorBGRA									color
 		, ::ced::container<::ced::SCoord2<int32_t>>			& pixelCoords
-		, ::ced::container<::ced::STriangleWeights<double>>	& pixelVertexWeights
+		, ::ced::container<::ced::STriangleWeights<float>>	& pixelVertexWeights
 		, ::ced::view_grid<uint32_t>						depthBuffer
 		);
 	int								drawTriangle
@@ -103,7 +103,7 @@ namespace ced
 		, const ::ced::SMatrix4<float>						& matrixTransformView
 		, const ::ced::SCoord3<float>						& lightVector
 		, ::ced::container<::ced::SCoord2<int32_t>>			& pixelCoords
-		, ::ced::container<::ced::STriangleWeights<double>>	& pixelVertexWeights
+		, ::ced::container<::ced::STriangleWeights<float>>	& pixelVertexWeights
 		, ::ced::view_grid<::ced::SColorBGRA>				textureImage
 		, ::ced::view_grid<uint32_t>						depthBuffer
 		);
@@ -116,7 +116,7 @@ namespace ced
 		, const ::ced::SCoord3<float>						& lightVector
 		, const ::ced::SColorFloat							& lightColor
 		, ::ced::container<::ced::SCoord2<int32_t>>			& pixelCoords
-		, ::ced::container<::ced::STriangleWeights<double>>	& pixelVertexWeights
+		, ::ced::container<::ced::STriangleWeights<float>>	& pixelVertexWeights
 		, ::ced::view_grid<::ced::SColorBGRA>				textureImage
 		, ::ced::container<::ced::SLight3>					& lightPoints
 		, ::ced::container<::ced::SColorBGRA>				& lightColors
