+++ ced_demo_06/ced_demo_06.cpp
@@ -86,7 +86,7 @@ int													update				(SApplication & app)	{
 	matrixView											= matrixView * matrixViewport;
 
 	::ced::container<::ced::SCoord2<int32_t>>				pixelCoords;
-	::ced::container<::ced::STriangleWeights<double>>		pixelVertexWeights;
+	::ced::container<::ced::STriangleWeights<float>>		pixelVertexWeights;
 	::ced::SModelMatrices									matrices;
 	for(uint32_t iModel = 0; iModel < app.Models.size(); ++iModel) {
 		matrices.Scale		.Scale			(app.Models[iModel].Scale	, true);
